--- a/server/memcached/src/main/java/org/infinispan/server/memcached/binary/BinaryHeader.java
+++ b/server/memcached/src/main/java/org/infinispan/server/memcached/binary/BinaryHeader.java
@@ -8,12 +8,13 @@ import org.infinispan.server.memcached.logging.Header;
  * @since 15.0
  **/
 public class BinaryHeader extends Header {
-   private final Object key;
-   final BinaryCommand op;
-   final int opaque;
-   long cas;
 
-   BinaryHeader(Temporal requestStart, int requestBytes, String principalName, Object key, BinaryCommand op, int opaque, long cas) {
+   private Object key;
+   private BinaryCommand op;
+   private int opaque;
+   private long cas;
+
+   private BinaryHeader(Temporal requestStart, int requestBytes, String principalName, Object key, BinaryCommand op, int opaque, long cas) {
       super(requestStart, requestBytes, principalName);
       this.key = key;
       this.op = op;
@@ -21,8 +22,8 @@ public class BinaryHeader extends Header {
       this.cas = cas;
    }
 
-   BinaryHeader(BinaryHeader h) {
-      this(h.requestStart, h.requestBytes, h.principalName, h.key, h.op, h.opaque, h.cas);
+   BinaryHeader() {
+      this(null, -1, null, null, null, -1, -1);
    }
 
    @Override
@@ -38,5 +39,33 @@ public class BinaryHeader extends Header {
    @Override
    public String getProtocol() {
       return "MCBIN";
+   }
+
+
+   public BinaryHeader replace(Temporal requestStart, int requestBytes, String principalName, Object key, BinaryCommand op, int opaque, long cas) {
+      this.requestStart = requestStart;
+      this.requestBytes = requestBytes;
+      this.principalName = principalName;
+      this.key = key;
+      this.op = op;
+      this.opaque = opaque;
+      this.cas = cas;
+      return this;
+   }
+
+   public BinaryCommand getCommand() {
+      return op;
+   }
+
+   public int getOpaque() {
+      return opaque;
+   }
+
+   public long getCas() {
+      return cas;
+   }
+
+   public void setCas(long cas) {
+      this.cas = cas;
    }
 }