… server/memcached/src/main/java/org/infinispan/server/memcached/MemcachedResponse.java
… server/memcached/src/main/java/org/infinispan/server/memcached/binary/BinaryAuthDecoder.java
… server/memcached/src/main/java/org/infinispan/server/memcached/binary/BinaryOpDecoder.java
… server/memcached/src/main/java/org/infinispan/server/memcached/logging/Log.java
… server/memcached/src/main/java/org/infinispan/server/memcached/logging/MemcachedAccessLogging.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TextAuthDecoder.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TextCommand.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TextDecoder.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TextHeader.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TextIntrinsics.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TextOpDecoder.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TextResponse.java
… server/memcached/src/main/java/org/infinispan/server/memcached/text/TokenReader.java
… server/memcached/src/main/resources/memcache-binary-auth.gr
… server/memcached/src/main/resources/memcache-binary-op.gr
… server/memcached/src/main/resources/memcache-text-auth.gr
… server/memcached/src/main/resources/memcache-text-op.gr
… server/memcached/src/test/java/org/infinispan/server/memcached/binary/MemcachedBinaryReplicationTest.java
… server/memcached/src/test/java/org/infinispan/server/memcached/test/MemcachedFunctionalTest.java
… server/memcached/src/test/java/org/infinispan/server/memcached/text/MemcachedStatsTest.java