https://issues.redhat.com/browse/ISPN-14813

Most of the changes are based on #10934 and #11006.

I ran locally using the `memtier_benchmark`. If anyone is interested in running, I ran with the configuration `memtier_benchmark -p 11222 -4 -P memcache_binary -d 9000 -R --pipeline=10 --test-time=60`, varying the protocol with `memcache_binary` and `memcache_text`, and running one time to warm up and another to collect the info, additionally:

```
4         Threads
50        Connections per thread
60        Seconds
```

Comparison with `--pipeline=1`: https://gist.github.com/jabolina/99ea2e4d8317f36ad12d445bc57672a9
Comparison with `--pipeline=10`: https://gist.github.com/jabolina/6db34b4e1afce874ff7247a9ea9625d3

We can see that `pipeline=1` performs slightly worse but performs better with `pipeline=10`.

I am opening for CI to run, and if anyone wants to take a look already. I am going to check allocations, although I expect a behavior similar to main.