@wburns 2023-06-23
We need to remove all the usages of Unpooled in this new code. RESP and Memcached were built on the idea of a single threaded server. Due to this they have consistency guarantees that multiple commands sent from the same client must be processed in order and with no overlapping. This is done in the RESP server by disabling auto read and not reading any more of the ByteBuf that was passed if a command was not completed synchronously. This allows us some optimizations with reusing a ByteBuf that we cannot do otherwise.