@jabolina 2023-06-27
Ok, several changes in this update:

* Not using `Unpooled` in binary;
* Binary allocates a single `BinaryHeader` per channel instead of per operation;
* Text allocates headers per operation if access logging is enabled;
* Text protocol parses the command ID without allocating a string;
* Text protocol allocates a single buffer per channel to parse the arguments instead of slicing the buffer;
* Update the handler for both protocols to reduce the lambda allocation for get and set operations;
* Added replicated tests for binary.

We could update both protocols to avoid allocating the response per operation and do something similar to the `BinaryHeader` with one instance per channel.