@jabolina 2023-06-27
Comparison using `memtier_benchmark` with the same setup as in the PR description.

<details>
  <summary>Binary pipeline 1 (-0.0168518%)</summary>
  
### MAIN:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec
----------------------------------------------------------------------------------------------------------------------------
Sets        23883.51          ---          ---         0.77878         0.73500         1.31100         6.39900    211007.34
Gets       238818.99       170.35    238648.64         0.75928         0.73500         1.29500         5.24700      9069.54
Waits           0.00          ---          ---             ---             ---             ---             ---          ---
Totals     262702.50       170.35    238648.64         0.76106         0.73500         1.29500         5.37500    220076.88
```

### NEW:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec
----------------------------------------------------------------------------------------------------------------------------
Sets        23879.42          ---          ---         0.77591         0.74300         1.27100         4.89500    210971.18
Gets       238778.81       166.68    238612.13         0.75971         0.73500         1.24700         4.03100      9068.01
Waits           0.00          ---          ---             ---             ---             ---             ---          ---
Totals     262658.23       166.68    238612.13         0.76118         0.73500         1.25500         4.07900    220039.19

```
</details>

<details>
  <summary>Text pipeline 1 (-1.32795%)</summary>
  
### MAIN:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
----------------------------------------------------------------------------------------------------------------------------
Sets        23602.17          ---          ---         0.78859         0.74300         1.31900         5.66300    208360.36
Gets       236002.96       165.46    235837.50         0.76814         0.74300         1.30300         4.60700      7425.86
Waits           0.00          ---          ---             ---             ---             ---             ---          ---
Totals     259605.13       165.46    235837.50         0.77000         0.74300         1.30300         4.70300    215786.22
```

### NEW:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec
----------------------------------------------------------------------------------------------------------------------------
Sets        23288.60          ---          ---         0.79703         0.75900         1.31900         5.79100    205592.19
Gets       232869.09       159.15    232709.95         0.77870         0.75100         1.29500         4.54300      7290.98
Waits           0.00          ---          ---             ---             ---             ---             ---          ---
Totals     256157.70       159.15    232709.95         0.78036         0.75100         1.29500         4.63900    212883.16
```
</details>

<details>
  <summary>Binary pipeline 10 (87.9454%)</summary>
  
### MAIN:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec
----------------------------------------------------------------------------------------------------------------------------
Sets        65746.46          ---          ---         2.76347         1.59900        19.96700        36.86300    580860.03 
Gets       657447.39      1343.73    656103.67         2.76130         1.59900        19.96700        36.86300     24967.74
Waits           0.00          ---          ---             ---             ---             ---             ---          ---
Totals     723193.85      1343.73    656103.67         2.76150         1.59900        19.96700        36.86300    605827.77
```

### NEW:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec
----------------------------------------------------------------------------------------------------------------------------
Sets       123566.07          ---          ---         1.46716         1.22300         7.42300        17.02300   1091687.56
Gets      1235643.75      5215.86   1230427.89         1.46805         1.22300         7.42300        17.02300     46925.78
Waits           0.00          ---          ---             ---             ---             ---             ---          --- 
Totals    1359209.82      5215.86   1230427.89         1.46797         1.22300         7.42300        17.02300   1138613.34
```
</details>

<details>
  <summary>Text pipeline 10 (92.6426%)</summary>
  
### MAIN:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec
----------------------------------------------------------------------------------------------------------------------------
Sets        62523.60          ---          ---         2.90514         1.74300        20.73500        37.37500    551959.20
Gets       625219.26      1216.34    624002.92         2.90338         1.74300        20.73500        37.11900     26534.81
Waits           0.00          ---          ---             ---             ---             ---             ---          ---
Totals     687742.86      1216.34    624002.92         2.90354         1.74300        20.73500        37.11900    578494.01
```

### NEW:
```text
============================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec
----------------------------------------------------------------------------------------------------------------------------
Sets       120445.63          ---          ---         1.50473         1.28700         6.52700        17.15100   1063295.58
Gets      1204440.05      4832.37   1199607.68         1.50559         1.28700         6.49500        17.15100     73072.51
Waits           0.00          ---          ---             ---             ---             ---             ---          ---
Totals    1324885.68      4832.37   1199607.68         1.50551         1.28700         6.49500        17.15100   1136368.09
```
</details>

The improvement I am comparing is the total ops/sec. The p99 has a smaller value in all executions with the changes.